DatasourcePicker: Include the number of datasources in reported event
