+++ public/app/features/datasources/components/picker/DataSourcePicker.tsx
@@ -12,13 +12,13 @@ import { DataSourceInstanceSettings, GrafanaTheme2 } from '@grafana/data';
 import { selectors } from '@grafana/e2e-selectors';
 import { Trans, t } from '@grafana/i18n';
 import { reportInteraction } from '@grafana/runtime';
-import { DataQuery, DataSourceRef } from '@grafana/schema';
+import { DataQuery, DataSourceJsonData, DataSourceRef } from '@grafana/schema';
 import { Button, Icon, Input, ModalsController, Portal, ScrollContainer, useStyles2 } from '@grafana/ui';
 import config from 'app/core/config';
 import { useKeyNavigationListener } from 'app/features/search/hooks/useSearchKeyboardSelection';
 import { defaultFileUploadQuery, GrafanaQuery } from 'app/plugins/datasource/grafana/types';
 
-import { useDatasource } from '../../hooks';
+import { useDatasource, useDatasources } from '../../hooks';
 
 import { DataSourceList } from './DataSourceList';
 import { DataSourceLogo, DataSourceLogoPlaceHolder } from './DataSourceLogo';
@@ -102,6 +102,18 @@ export function DataSourcePicker(props: DataSourcePickerProps) {
   const currentValue = Boolean(!current && noDefault) ? undefined : currentDataSourceInstanceSettings;
   const prefixIcon =
     filterTerm && isOpen ? <DataSourceLogoPlaceHolder /> : <DataSourceLogo dataSource={currentValue} />;
+  const dataSources = useDatasources({
+    alerting: props.alerting,
+    annotations: props.annotations,
+    dashboard: props.dashboard,
+    logs: props.logs,
+    metrics: props.metrics,
+    mixed: props.mixed,
+    pluginId: props.pluginId,
+    tracing: props.tracing,
+    type: props.type,
+    variables: props.variables,
+  });
 
   // the order of middleware is important!
   const middleware = [
@@ -235,6 +247,7 @@ export function DataSourcePicker(props: DataSourcePickerProps) {
             item: INTERACTION_ITEM.OPEN_DROPDOWN,
             creator_team: 'grafana_plugins_catalog',
             schema_version: '1.0.0',
+            total_configured: dataSources.length,
           });
         }}
       >
@@ -282,13 +295,18 @@ export function DataSourcePicker(props: DataSourcePickerProps) {
                 onClose();
                 if (ds.uid !== currentValue?.uid) {
                   onChange(ds, defaultQueries);
-                  reportInteraction(INTERACTION_EVENT_NAME, { item: INTERACTION_ITEM.SELECT_DS, ds_type: ds.type });
+                  reportInteraction(INTERACTION_EVENT_NAME, {
+                    item: INTERACTION_ITEM.SELECT_DS,
+                    ds_type: ds.type,
+                    total_configured: dataSources.length,
+                  });
                 }
               }}
               onClose={onClose}
               onClickAddCSV={onClickAddCSV}
               onDismiss={onClose}
               onNavigateOutsiteFooter={onNavigateOutsiteFooter}
+              dataSources={dataSources}
             />
           </div>
         </Portal>
@@ -325,10 +343,11 @@ export interface PickerContentProps extends DataSourcePickerProps {
   onDismiss: () => void;
   footerRef: (element: HTMLElement | null) => void;
   onNavigateOutsiteFooter: (e: React.KeyboardEvent<HTMLButtonElement>) => void;
+  dataSources: Array<DataSourceInstanceSettings<DataSourceJsonData>>;
 }
 
 const PickerContent = React.forwardRef<HTMLDivElement, PickerContentProps>((props, ref) => {
-  const { filterTerm, onChange, onClose, onClickAddCSV, current, filter } = props;
+  const { filterTerm, onChange, onClose, onClickAddCSV, current, filter, dataSources } = props;
 
   const changeCallback = useCallback(
     (ds: DataSourceInstanceSettings) => {
@@ -360,6 +379,7 @@ const PickerContent = React.forwardRef<HTMLDivElement, PickerContentProps>((prop
               item: INTERACTION_ITEM.CONFIG_NEW_DS_EMPTY_STATE,
             })
           }
+          dataSources={dataSources}
         ></DataSourceList>
       </ScrollContainer>
       <FocusScope>
